scripts/prospecting/requirements.txt
# Dependencies for scripts/prospecting/ agents
# Pin versions for reproducible builds in GitHub Actions

anthropic>=0.40.0,<1.0.0
google-generativeai>=0.8.0,<1.0.0
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
aiohttp>=3.10.0,<4.0.0
